% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{add_facet_lines}
\alias{add_facet_lines}
\title{Add axis lines to facetted plots}
\usage{
add_facet_lines(size = 1)
}
\arguments{
\item{size}{Line size (in mm).}
}
\description{
Draw axis lines to facetted plots.
}
