% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_posterior_density.R
\name{add_posterior_density}
\alias{add_posterior_density}
\title{Compute the Kernel density estimates and highest density interval of
posterior distributions}
\usage{
add_posterior_density(data, estimates, prob, outer_prob = 0.99)
}
\arguments{
\item{data}{A data frame.}

\item{estimates}{Variable containing posterior distributions.}

\item{prob}{The probability mass to include in the shaded region.}

\item{outer_prob}{The probability mass to include in the outer interval.}
}
\description{
Add the Kernel density estimates and highest density interval of
posterior distributions to a data frame prior to using \code{\link[=posterior_dist]{posterior_dist()}}.
}
