% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pred_conf.R
\name{add_pred_conf}
\alias{add_pred_conf}
\title{Add predictions from the posterior fit to a data frame}
\usage{
add_pred_conf(x, fit, formula)
}
\arguments{
\item{x}{A data frame.}

\item{fit}{A JAGS model fit.}

\item{formula}{Formula containing group-level effects to be considered in the prediction.}
}
\description{
Add the posterior predictive with standard deviation and confidence from a JAGS model to a data frame.
}
