% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_draws.R
\name{add_quantiles}
\alias{add_quantiles}
\title{Add the lowest and highest quantile values to a data frame}
\usage{
add_quantiles(x, var, prob = 0.95)
}
\arguments{
\item{x}{A data frame.}

\item{var}{Column of numeric values to compute quantiles from.}

\item{prob}{The probability mass to get the lowest and highest quantiles from.}
}
\description{
Add the lowest and highest quantile values to a data frame.
}
