% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_max_diff}
\alias{calc_max_diff}
\title{Calculate the maximum difference}
\usage{
calc_max_diff(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Calculate the difference between the two most extreme values in
a numeric vector.
}
