% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistical_details.R
\name{calc_prob_inf_0}
\alias{calc_prob_inf_0}
\title{Calculate the probability of posterior fits to be lower than 0}
\usage{
calc_prob_inf_0(data)
}
\arguments{
\item{x}{A data frame containing posterior fits.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
Calculate the probability of posterior fits to be lower than 0.
}
