% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_mdl_data.R
\name{compose_mdl_data}
\alias{compose_mdl_data}
\title{Compose JAGS-compatible data}
\usage{
compose_mdl_data(data, matrix_formula = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{matrix_formula}{A formula.}
}
\value{
A list of data to pass on to a JAGS model.
}
\description{
Turn a data frame into a list of data to pass on to a JAGS model.
}
