% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_jags_r2.R
\name{compute_jags_r2}
\alias{compute_jags_r2}
\title{Compute Bayesian R-squared}
\usage{
compute_jags_r2(x, formula)
}
\arguments{
\item{x}{A model object file returned by \code{\link[=run_jags]{run_jags()}}.}

\item{formula}{Formula containing group-level effects to be considered in the prediction.}
}
\description{
Compute Bayesian R-squared from a JAGS model output.
}
