% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{delta}
\alias{delta}
\title{Compute the difference between consecutive values}
\usage{
delta(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Compute the difference between the first element of a numeric vector and all consecutive values.
}
