% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_models.R
\name{eval_models}
\alias{eval_models}
\title{Evaluate JAGS models}
\usage{
eval_models(files)
}
\arguments{
\item{files}{Path of files containing a JAGS model object returned by \code{\link[=run_jags]{run_jags()}}.}
}
\value{
A tibble of WAIC and LOO IC estimates with standard error.
}
\description{
Compare multiple JAGS models using the WAIC and leave-one-out (LOO) cross-validation methods.
}
