% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_jags_summary.R
\name{extend_jags_summary}
\alias{extend_jags_summary}
\title{Extend JAGS summary}
\usage{
extend_jags_summary(data)
}
\arguments{
\item{data}{A data frame.}
}
\description{
Add standardized slope estimates and whether the 95\\% credible
interval (CI) includes 0 or not to a JAGS summary. A value of 1 indicates
that the CI includes 0, a value of 0 indicates that the CI does not include
0. Requires running \code{\link[=run_jags]{run_jags()}} with \code{default_output=FALSE}.
}
