% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{extract_file_name}
\alias{extract_file_name}
\title{Extract file names}
\usage{
extract_file_name(path, to_snake_case = FALSE)
}
\arguments{
\item{path}{A character vector of path names.}

\item{to_snake_case}{Should the file name be converted to snake case?}
}
\description{
Extract the file name from a full path name.
}
