% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{first_word}
\alias{first_word}
\title{Extract the first word in a character string}
\usage{
first_word(string, compound_word = TRUE)
}
\arguments{
\item{string}{A character vector.}

\item{compound_word}{If \code{TRUE}, also captures hyphenated compound word.}
}
\value{
A character vector.
}
\description{
Extract the first word in a character string.
}
