% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_files.R
\name{get_files}
\alias{get_files}
\title{Fetch files}
\usage{
get_files(
  dir,
  vars = NULL,
  elevation_span = NULL,
  exclusion_zone = NULL,
  std_from = c("top", "bottom", "none"),
  land_type = FALSE,
  ...
)
}
\arguments{
\item{dir}{Directory fetch files from.}

\item{vars}{Explanatory variables to use.}

\item{elevation_span}{Elevation spans to use.}

\item{exclusion_zone}{Exclusion zones to use.}

\item{std_from}{Side of standardization to use. One of \code{top}, \code{bottom} or
\code{none}.}

\item{land_type}{If \code{TRUE}, select files from analyses performed by land type.}

\item{...}{Other arguments passed to \code{\link[=list_files]{list_files()}}.}
}
\value{
A vector of file dirs with class \code{collection}.
}
\description{
Fetch files produced by \code{\link[=run_jags]{run_jags()}}.
}
\examples{
\dontrun{
get_files(
  "path_to_dir",
  vars = c("dtr", "ts"),
  elevation_span = 2000
)
}
}
