% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_sims.R
\name{get_jags_sims}
\alias{get_jags_sims}
\title{Extract \code{sims} data from a JAGS object}
\usage{
get_jags_sims(x, vars)
}
\arguments{
\item{x}{A data frame.}

\item{vars}{Variables to select.}
}
\description{
Extract \code{sims} data from a model object returned by \code{\link[=run_jags]{run_jags()}}. Requires running \code{\link[=run_jags]{run_jags()}} with \code{save=TRUE} and \code{default_output=FALSE}.
}
