% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistical_details.R
\name{get_statistical_details}
\alias{get_statistical_details}
\title{Compute Bayesian R-squared and the probability of posterior fits to be lower than 0}
\usage{
get_statistical_details(x)
}
\arguments{
\item{x}{A model object file returned by \code{\link[=run_jags]{run_jags()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}.
}
\description{
Compute Bayesian R-squared and the probability of posterior fits to be lower than 0 from a JAGS model output.
}
