% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closures.R
\name{global_local_analyses}
\alias{global_local_analyses}
\alias{global_analyses}
\alias{local_analyses}
\title{Global- and local-scale analyses}
\usage{
global_analyses(path)

local_analyses(path)
}
\arguments{
\item{path}{Path to a directory containing model objects returned by \code{\link[=run_jags]{run_jags()}}.}
}
\description{
Closures to plot model estimates and get model statistics.
}
