% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_metachr}
\alias{has_metachr}
\title{Check for metacharacters}
\usage{
has_metachr(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A boolean.
}
\description{
Test whether a string contains metacharacters or not.
}
