% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{line_0}
\alias{line_0}
\title{Draw a line at 0}
\usage{
line_0(
  axis = c("x", "y"),
  linetype = "dashed",
  color = "grey50",
  size = 0.3,
  ...
)
}
\arguments{
\item{axis}{Axis to draw the line on.}

\item{linetype}{Line type.}

\item{color}{Line color.}

\item{size}{Line size (in mm).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}.}
}
\description{
Draw a vertical or horizontal line at 0.
}
