% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List files in a directory}
\usage{
list_files(path = ".", target = NULL, names, full.names = TRUE, ...)
}
\arguments{
\item{path}{A character vector of full path names.}

\item{target}{An optional \code{\link[=regex]{regular expression}} to return only file
names that match the regular expression.}

\item{names}{A vector of names to assign to each file, a function, formula or
regular expression to extract names from file names.}

\item{full.names}{Should the full path be returned?}

\item{...}{Other arguments passed on to \code{\link[=list.files]{list.files()}}.}
}
\value{
A character vector of class \code{collection}.
}
\description{
A wrapper around \code{\link[=list.files]{list.files()}} that gives the possibility to
name each file directly.
}
