% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mdl.R
\name{load_mdl}
\alias{load_mdl}
\title{Load a JAGS model}
\usage{
load_mdl(file)
}
\arguments{
\item{file}{File containing a JAGS model.}
}
\value{
A list of model-related data that includes the name of the model, parameters to save, file path of the model and the actual model to run.
}
\description{
Load a file containing a JAGS model and parameters.
}
