% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_model}
\alias{load_model}
\title{Load model}
\usage{
load_model(scope = c("global", "local"))
}
\arguments{
\item{scope}{Scope of the model to load. One of \code{c("global", "local")}.}
}
\description{
Load the global- or local-analyses model file.
}
