% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_dir}
\alias{make_dir}
\title{Parse numeric values from a string}
\usage{
make_dir(path, ...)
}
\arguments{
\item{path}{A character vector containing a single path name.}

\item{...}{Other arguments passed on to \code{\link[=dir.create]{dir.create()}}.}
}
\description{
Wrapper around \code{\link[=dir.create]{dir.create()}} that creates directories recursively.
}
