% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_list}
\alias{make_list}
\title{Create named lists dynamically}
\usage{
make_list(...)
}
\arguments{
\item{...}{Objects to pass on to \code{\link[rlang:list2]{dots_list()}}.}
}
\description{
Wrapper around \code{\link[rlang:list2]{dots_list()}} with
\code{.named = TRUE} as default.
}
