% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{make_posterior_data}
\alias{make_posterior_data}
\title{Prepare data for plotting posterior distribution}
\usage{
make_posterior_data(
  files,
  yvar,
  prob = c(0.8, 0.95),
  prob_outer = 0.99,
  scales = 0.01,
  labels = NULL,
  reverse = FALSE
)
}
\arguments{
\item{files}{Files to get data from.}

\item{yvar}{Variable describing the vertical axis.}

\item{prob}{The probability mass to include in the shaded region.}

\item{prob_outer}{The probability mass to include in the outer interval.}

\item{scales}{A scaling factor to scale the height of the ridgelines. See
\code{\link[ggridges:geom_ridgeline]{ggridges::geom_ridgeline()}} for details.}

\item{labels}{A named vector of labels used for plotting. Names must be the
same as the explanatory variables.}

\item{reverse}{Should the order of posterior distributions be reversed?}
}
\description{
Structure data for \code{\link[=plot_posterior_distributions]{plot_posterior_distributions()}}.
}
\examples{
\dontrun{
get_files("path_to_dir", vars = c("dtr", "ts"), elevation_span = 2000) |>
  make_posterior_data(yvar = "exclusion_zone")
}
}
