% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{make_regression_data}
\alias{make_regression_data}
\title{Prepare data for regression plots}
\usage{
make_regression_data(files, by_land_types = FALSE)
}
\arguments{
\item{files}{Files to get data from.}

\item{by_land_types}{Should estimates be calculated by land type?}
}
\description{
Structure data for \code{\link[=regressions]{regressions()}}.
}
