% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_num}
\alias{parse_num}
\title{Parse numeric values from a string}
\usage{
parse_num(x, as = c("numeric", "integer", "character"), all = FALSE)
}
\arguments{
\item{x}{A character string.}

\item{as}{A character vector describing the type of output values. One of
\code{numeric}, \code{integer} or \code{character}.}

\item{all}{Should all numeric values be extracted? By default, parses the first
numeric value.}
}
\description{
Extract numeric values from a character string.
}
