% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regressions.R
\name{plot_regressions}
\alias{plot_regressions}
\title{Plot regression fits}
\usage{
plot_regressions(
  x,
  y,
  facet_by,
  show_draws = TRUE,
  n_draws = 600,
  draws_prob = 0.95,
  draws_color = "#b6d3e2",
  draws_size = 0.25,
  draws_alpha = 0.1,
  mean_draw_color = "#006699",
  point_colors = c("#737373", "#ffffff"),
  land_type = FALSE,
  ribbon_colors = NULL,
  error_bars = TRUE,
  point_labels = FALSE,
  label_colors = NULL,
  labellers,
  legend_position = c("none", "right", "bottom", "top"),
  text_size = 11,
  panel_spacing = unit(0.5, "lines"),
  seed = 130821
)
}
\arguments{
\item{x}{A data frame of observed data.}

\item{y}{A data frame of model fits.}

\item{facet_by}{Variable defining faceting groups.}

\item{show_draws}{Should estimated draws be shown?}

\item{n_draws}{Number of draws to sample from the 95\\\% credible interval.}

\item{draws_prob}{Probability mass to sample the draws from.}

\item{draws_color}{Color of the draws.}

\item{draws_size}{Size of the draws.}

\item{draws_alpha}{Transparency the of draws.}

\item{mean_draw_color}{Color of the mean draw.}

\item{point_colors}{Color of the points data.}

\item{land_type}{If \code{TRUE}, will plot regression lines with the 95\\\% credible interval for each land type.}

\item{ribbon_colors}{Color of the ribbons. Only applies when \code{land_type=TRUE}.}

\item{error_bars}{Should error bars be drawn?}

\item{point_labels}{Should point labels be shown?}

\item{label_colors}{Color of the labels.}

\item{labellers}{A named character vector of labellers for the x axis.}

\item{legend_position}{Position of the legend. One of \code{c("right", "bottom")}.}

\item{text_size}{A numerical value defining the text size (in pts).}

\item{panel_spacing}{Spacing between facet panels. See \code{\link[ggplot2:theme]{ggplot2::theme()}} for details.}

\item{seed}{Seed value for draw sampling.}
}
\description{
Plot regression fits.
}
