% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{posterior_dist}
\alias{posterior_dist}
\title{Plot posterior distribution from a bayesian model}
\usage{
posterior_dist(
  x,
  mapping = aes(),
  scale = 1,
  vline_type = 1,
  vline_color = "grey90",
  facet_args = list(),
  ...
)
}
\arguments{
\item{x}{A data frame.}

\item{mapping}{Default list of aesthetic mappings to use for plot. See \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} for details.}

\item{scale}{A scaling factor to scale the height of the ridgelines. See \code{\link[ggridges:geom_ridgeline]{ggridges::geom_ridgeline()}} for details.}

\item{vline_type}{Line type of the line indicating 0.}

\item{vline_color}{Color of the line indicating 0.}

\item{facet_args}{A named list of parameters and arguments to pass on to \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}
}
\description{
Plot posterior distribution from a bayesian model.
}
