% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ppc}
\alias{ppc}
\title{Posterior predictive checks}
\usage{
ppc(y, yrep, n_draws = 100, ...)
}
\arguments{
\item{y}{Observed data.}

\item{yrep}{Fitted data.}

\item{n_draws}{Number of fitted draws to sample.}

\item{...}{Optional arguments passed on to \code{\link[bayesplot:PPC-distributions]{bayesplot::ppc_dens_overlay()}}.}
}
\description{
Wrapper around \code{\link[bayesplot:PPC-distributions]{bayesplot::ppc_dens_overlay()}} to plot posterior
predictive check from a JAGS model object.
}
