% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{ppc_}
\alias{ppc_}
\title{Posterior predictive checks}
\usage{
ppc_(y, yrep, all = FALSE, ...)
}
\arguments{
\item{y}{Observed data.}

\item{yrep}{Fitted data.}

\item{all}{If \code{TRUE}, will use all species ranges passed to \code{\link[=run_jags]{run_jags()}}. Otherwise, will use species ranges averaged in each location.}

\item{...}{Optional arguments passed on to \code{\link[bayesplot:PPC-distributions]{bayesplot::ppc_dens_overlay()}}.}
}
\description{
More specific variant of \code{\link[=ppc]{ppc()}} that will check posteriors from the global-scale analysis.
}
