% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{proportion}
\alias{proportion}
\title{Compute proportion}
\usage{
proportion(condition)
}
\arguments{
\item{condition}{Logical vectors.}
}
\description{
Context dependent expression that returns the current proportion
of values that sastify a condition. See \code{\link[dplyr:context]{dplyr::context()}} for details.
}
