% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jags.R
\name{read_jags}
\alias{read_jags}
\alias{read_jags.character}
\alias{read_jags.collection}
\title{Read JAGS custom output files}
\usage{
read_jags(x, pluck, vars = NULL, ...)

\method{read_jags}{character}(x, pluck, vars = NULL, ...)

\method{read_jags}{collection}(x, pluck, vars = NULL, ..., names_to = NULL)
}
\arguments{
\item{x}{A \code{.rds} file or a collection of \code{.rds} files.}

\item{pluck}{Object to be extracted from a JAGS output.}

\item{vars}{Columns to select or model parameters to filter from the \code{sims}
or \code{summary} data, respectively.}

\item{...}{Arguments passed on to methods.}

\item{names_to}{Either a string or \code{NULL}. See \code{\link[purrr:list_c]{purrr::list_rbind()}} for
details.}
}
\description{
Read model object files returned by \code{\link[=run_jags]{run_jags()}}. Requires
running \code{run_jags()} with \code{save=TRUE} and \code{default_output = FALSE}.
}
