% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{regressions}
\alias{regressions}
\alias{regressions.draws}
\alias{regressions.land_types}
\title{Plot regression fits}
\usage{
regressions(data, facet_by = "expl_var", labels = NULL, ...)

\method{regressions}{draws}(
  data,
  facet_by = "expl_var",
  labels = NULL,
  ...,
  n_draws = 600,
  draws_prob = 0.95,
  point_labels = FALSE,
  seed = 130821
)

\method{regressions}{land_types}(data, facet_by = "expl_var", labels = NULL, ...)
}
\arguments{
\item{data}{A list of data produced by \code{\link[=make_regression_data]{make_regression_data()}}.}

\item{facet_by}{Variable defining faceting groups.}

\item{labels}{A character vector used to label the x axis. If \code{labels} contains
several labels, each label must be named. Names must be the same as the
explanatory variables.}

\item{...}{Other arguments passed to methods.}

\item{n_draws}{Number of draws to sample from the 95\\\% credible interval.}

\item{draws_prob}{Probability mass to sample the draws from.}

\item{point_labels}{Should point labels be shown?}

\item{seed}{Seed value for draw sampling.}
}
\description{
Plot regression fits.
}
