% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_nearest}
\alias{round_nearest}
\title{Round up or down a number to the nearest value}
\usage{
round_nearest(x, nearest = -10)
}
\arguments{
\item{x}{A numeric vector.}

\item{nearest}{A numeric value indicating the nearest value to round by. If
negative, will round down to the nearest value. If positive, will round up
to the nearest value.}
}
\description{
Round up or down a number to the nearest value.
}
