% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_crop}
\alias{rs_crop}
\title{Crop a raster object}
\usage{
rs_crop(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{\link[terra:crop]{terra::crop()}}.}
}
\description{
Alias for \code{\link[terra:crop]{terra::crop}}.
}
