% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_ext_to_polygon}
\alias{rs_ext_to_polygon}
\title{Convert a raster extent to a polygon}
\usage{
rs_ext_to_polygon(x)
}
\arguments{
\item{x}{A raster object.}
}
\description{
Convert a raster extent to a polygon.
}
