% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_extent}
\alias{rs_extent}
\title{Get or set the extent of a raster object}
\usage{
rs_extent(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{\link[terra:ext]{terra::ext()}}.}
}
\description{
Alias for \code{\link[terra:ext]{terra::ext}}.
}
