% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_extract}
\alias{rs_extract}
\title{Extract values from a raster object}
\usage{
rs_extract(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{\link[terra:extract]{terra::extract()}}.}
}
\description{
Alias for \code{\link[terra:extract]{terra::extract}}.
}
