% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_project}
\alias{rs_project}
\title{Reproject a spatial object}
\usage{
rs_project(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{\link[terra:project]{terra::project()}}.}
}
\description{
Alias for \code{\link[terra:project]{terra::project}}.
}
