% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_read}
\alias{rs_read}
\title{Read a raster object}
\usage{
rs_read(x, ...)
}
\arguments{
\item{x}{A single or collection of file names.}

\item{...}{Arguments to pass on to \code{\link[terra:rast]{terra::rast()}}.}
}
\description{
Alias for \code{\link[terra:rast]{terra::rast}}.
}
