% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_read_stk}
\alias{rs_read_stk}
\title{Read a multi-layer raster file}
\usage{
rs_read_stk(file, layer_names = NULL)
}
\arguments{
\item{file}{A file containing a raster stack.}

\item{layer_names}{A character vector containing the names to assign to each raster layer.}
}
\description{
Read a multi-layer raster file.
}
