% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_resample}
\alias{rs_resample}
\title{Resample a spatial object}
\usage{
rs_resample(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{\link[terra:resample]{terra::resample()}}.}
}
\description{
Alias for \code{\link[terra:resample]{terra::resample}}.
}
