% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_set_crs}
\alias{rs_set_crs}
\title{Set the coordinate reference system of a raster object}
\usage{
rs_set_crs(x, proj = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{x}{A aster object.}

\item{proj}{Projection to set to the raster object.}
}
\description{
Set the coordinate reference system of a raster object.
}
