% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_set_range}
\alias{rs_set_range}
\title{Set the minimum and maximum values of a raster object}
\usage{
rs_set_range(x)
}
\arguments{
\item{x}{A raster object.}
}
\description{
Remove missing values and set the minimum and maximum values of
a raster object.
}
