% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_vect}
\alias{rs_vect}
\title{Create spatial vector objects}
\usage{
rs_vect(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{\link[terra:vect]{terra::vect()}}.}
}
\description{
Alias for \code{\link[terra:vect]{terra::vect}}.
}
