% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_zonal}
\alias{rs_zonal}
\title{Compute zonal statistics}
\usage{
rs_zonal(...)
}
\arguments{
\item{...}{Arguments to pass on to \code{\link[terra:zonal]{terra::zonal()}}.}
}
\description{
Alias for \code{\link[terra:zonal]{terra::zonal}}.
}
