% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_jags.R
\name{run_jags}
\alias{run_jags}
\title{Run a JAGS model}
\usage{
run_jags(
  x,
  model,
  ...,
  save = FALSE,
  path,
  filename = NULL,
  default_output = FALSE
)
}
\arguments{
\item{x}{Data to pass on to a JAGS model. See \code{\link[R2jags:jags]{R2jags::jags()}} for details.}

\item{model}{List containing a JAGS model and parameters.}

\item{...}{Other arguments to pass on to \code{\link[R2jags:jags]{R2jags::jags()}}.}

\item{save}{Should the JAGS output be saved?}

\item{path}{Directory to save the file in.}

\item{filename}{Name (without extension) of the file. Leave \code{filename=NULL} to use a default file name based on the model settings.}

\item{default_output}{Should \code{run_jags()} return the model object returned by \code{\link[R2jags:jags]{R2jags::jags()}}?}
}
\description{
Run a JAGS model.
}
