% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gis.R
\name{setup_gis}
\alias{setup_gis}
\title{Create directories for GIS analyses}
\usage{
setup_gis()
}
\description{
Create the required directories to run the GIS analyses in
\link{Diurnal temperature range as a key predictor of plants' elevation ranges globally}
(https://github.com/arnaudgallou/TRS)
}
