% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_draws.R
\name{slice_draws}
\alias{slice_draws}
\title{Subset rows from a credible interval}
\usage{
slice_draws(x, estimates, prob = 0.95, n = 200, seed)
}
\arguments{
\item{x}{A data frame.}

\item{estimates}{Column to subset the draws from.}

\item{prob}{The probability mass to subset.}

\item{n}{Number of draws to subset.}

\item{seed}{Random \code{\link[=set.seed]{seed}} to produce identical result.}
}
\description{
Wrapper around the \code{\link[dplyr:slice]{dplyr::slice_sample()}} that subsets rows from a credible interval.
}
\examples{
\dontrun{
df <- tibble(id = 1:500, value = rnorm(n = 500, mean = 0, sd = 5))
slice_draws(df, estimates = value)
}
}
