% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string_clean}
\alias{string_clean}
\title{Remove non-letter characters, underscore and repeated whitespace inside
a string}
\usage{
string_clean(string)
}
\arguments{
\item{string}{A character vector.}
}
\value{
A character vector.
}
\description{
Remove any non-letter characters from the beginning and end of a
character string as well as underscores and repeated whitespaces inside a
string.
}
