% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string_extract}
\alias{string_extract}
\alias{string_extract_all}
\title{Replace matching patterns in a string}
\usage{
string_extract(string, pattern)

string_extract_all(string, pattern)
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{A pattern to look for.}
}
\value{
A character vector.

A character vector.
}
\description{
Vectorised extraction of matching patterns using Perl-compatible
regular expressions.
}
