% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{string_replace}
\alias{string_replace}
\alias{string_replace_all}
\title{Replace matched patterns in a string}
\usage{
string_replace(string, pattern, replacement)

string_replace_all(string, pattern, ignore_case = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{pattern}{Pattern to look for. A single element vector if \code{string_replace}
or a named vector of patterns and replacements if \code{string_replace_all}.}

\item{replacement}{A character vector of replacement. Only applies to
\code{string_replace}.}

\item{ignore_case}{Should letter case be ignored? Only applies to \code{string_replace_all}.}
}
\value{
A character vector.
}
\description{
Vectorised pattern replacement using Perl-compatible regular
expressions.
}
