% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_jags_summary.R
\name{summarize_slopes}
\alias{summarize_slopes}
\title{Compute the proportion of positive and negative slopes}
\usage{
summarize_slopes(x)
}
\arguments{
\item{x}{A data frame.}
}
\description{
Compute the proportion of positive and negative slopes in a data
frame returned by \code{\link[=extend_jags_summary]{extend_jags_summary()}}. High and low uncertainties
indicate that the 95\\% CI includes 0 or not, respectively.
}
