% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\title{Add a tag to a plot}
\usage{
tag(
  tag,
  position = c("topleft", "topright", "bottomleft", "bottomright"),
  margin,
  fontface = 2,
  geom = "text",
  ...
)
}
\arguments{
\item{tag}{Tag to display.}

\item{position}{Position of the tag. One of \code{topleft}, \code{topright}, \code{bottomleft},
\code{bottomright}.}

\item{margin}{Numeric vector indicating the space between the tag element and
border of the plot region. If two values are passed, the first value defines
the horizontal margin while the second value defines the vertical margin.}

\item{fontface}{Font face of the tag element.}

\item{geom}{Name of geom to use for annotation.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:annotate]{annotate()}}.}
}
\description{
Add a tag inside the plot region.
}
