% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\alias{theme_elesic}
\alias{theme_blank_y}
\title{Custom ggplot themes}
\usage{
theme_elesic(
  base_size = 11,
  base_family = "",
  base_line_size = 0.5,
  base_rect_size = base_size/22,
  base_axis_color = "black",
  legend_position = "none"
)

theme_blank_y(
  base_size = 11,
  base_family = "",
  base_line_size = 0.5,
  base_rect_size = base_size/22,
  base_axis_color = "black",
  legend_position = "none"
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{base_line_size}{Base size for line elements.}

\item{base_rect_size}{Base size for rect elements.}

\item{base_axis_color}{Base color for axis elements. Applies to \code{axis.line},
\code{axis.ticks} and \code{axis.text}.}

\item{legend_position}{Set the legend position. See \code{\link[ggplot2:theme]{ggplot2::theme()}} for
details. Set to \code{"none"} to hide the legend.}
}
\description{
Custom ggplot themes.
}
