% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{to_snake_case}
\alias{to_snake_case}
\title{To snake case}
\usage{
to_snake_case(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector.
}
\description{
Convert a character string to snake case.
}
\examples{
to_snake_case("laCigaleEtLaFourmi")
}
