% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs.utilities-package.R
\docType{package}
\name{TRS.utilities-package}
\alias{TRS.utilities}
\alias{TRS.utilities-package}
\title{TRS.utilities: TRS Utility Functions}
\description{
Helpers functions to reproduce results from the TRS project.
}
\keyword{internal}
