% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{uc_first}
\alias{uc_first}
\title{Convert a string's first character to uppercase}
\usage{
uc_first(string, to_lower = FALSE)
}
\arguments{
\item{string}{A character vector.}

\item{to_lower}{Should other characters be converted to lowercase?}
}
\value{
A character vector.
}
\description{
Convert a string's first character to uppercase.
}
