% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\docType{data}
\name{WGS84}
\alias{WGS84}
\title{WGS84 coordinate reference system}
\format{
An object of class \code{character} of length 1.
}
\usage{
WGS84
}
\description{
WGS84 coordinate reference system.
}
\keyword{datasets}
