% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_statistical_details.R
\name{calc_p_inf_0}
\alias{calc_p_inf_0}
\title{Calculate the probability of posterior fits to be lower than 0}
\usage{
calc_p_inf_0(x)
}
\arguments{
\item{x}{A model object file returned by \code{\link[=run_jags]{run_jags()}}.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}}.
}
\description{
Calculate the probability of posterior fits to be lower than 0.
}
