% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-local-analyses.R
\name{global_local_analyses}
\alias{global_local_analyses}
\alias{GlobalAnalyses}
\alias{LocalAnalyses}
\title{Global- and local-scale analyses}
\description{
Classes to plot model estimates and get model statistics.
}
\examples{
\dontrun{

# Global scale analyses
ga <- GlobalAnalyses$new(path)

ga$regressions(vars = "dtr", elev_span = 2500, excl_zone = 250, labels = "DTR")

ga$posterior_distributions(vars = c("dtr", "ts"))

# Local scale analyses
la <- LocalAnalyses(path)

la$slope_histograms()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GlobalAnalyses-new}{\code{GlobalAnalyses$new()}}
\item \href{#method-GlobalAnalyses-regressions}{\code{GlobalAnalyses$regressions()}}
\item \href{#method-GlobalAnalyses-posterior_distributions}{\code{GlobalAnalyses$posterior_distributions()}}
\item \href{#method-GlobalAnalyses-eval_models}{\code{GlobalAnalyses$eval_models()}}
\item \href{#method-GlobalAnalyses-get_statistical_details}{\code{GlobalAnalyses$get_statistical_details()}}
\item \href{#method-GlobalAnalyses-clone}{\code{GlobalAnalyses$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalAnalyses-new"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalAnalyses-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalAnalyses$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to a directory containing the outputs of the global-scale
analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalAnalyses-regressions"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalAnalyses-regressions}{}}}
\subsection{Method \code{regressions()}}{
Plot regression fits.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalAnalyses$regressions(
  vars,
  elev_span = NULL,
  excl_zone = NULL,
  std_from = c("top", "bottom", "none"),
  facet_by = "expl_var",
  labels = NULL,
  by_land_type = FALSE,
  point_labels = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{A character vector of variables to plot.}

\item{\code{elev_span}}{Length of elevation spans to use.}

\item{\code{excl_zone}}{Length of exclusion zones to use.}

\item{\code{std_from}}{Edge of elevational gradients to use. One of \code{top}, \code{bottom},
\code{none}. Use \code{none} for data that were not standardized.}

\item{\code{facet_by}}{Variable defining faceting groups.}

\item{\code{labels}}{A character vector used to label the x axis. See
\code{\link[=regressions]{regressions()}} for details.}

\item{\code{by_land_type}}{Should regressions be drawn for each land type?}

\item{\code{point_labels}}{Should point labels be shown?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalAnalyses-posterior_distributions"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalAnalyses-posterior_distributions}{}}}
\subsection{Method \code{posterior_distributions()}}{
Plot posterior distribution from a bayesian model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalAnalyses$posterior_distributions(
  vars,
  elev_span = NULL,
  excl_zone = NULL,
  std_from = c("top", "bottom", "none"),
  yvar = "exclusion_zone",
  prob = c(0.8, 0.95),
  prob_outer = 0.99,
  scales = 0.01,
  labels = NULL,
  reverse = FALSE,
  facet = TRUE,
  fill = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{A character vector of variables to plot.}

\item{\code{elev_span}}{Length of elevation spans to use.}

\item{\code{excl_zone}}{Length of exclusion zones to use.}

\item{\code{std_from}}{Edge of elevational gradients to use. One of \code{top}, \code{bottom},
\code{none}. Use \code{none} for data that were not standardized.}

\item{\code{yvar}}{Variable describing the vertical axis.}

\item{\code{prob}}{The probability mass to include in the shaded region.}

\item{\code{prob_outer}}{The probability mass to include in the outer interval.}

\item{\code{scales}}{A scaling factor to scale the height of the ridgelines. See
[\code{ggridges::geom_ridgeline()}][] for details.}

\item{\code{labels}}{A character vector used to label the x axis. See
\code{\link[=posterior_distributions]{posterior_distributions()}} for details.}

\item{\code{reverse}}{Should the order of posterior distributions be reversed?}

\item{\code{facet}}{Should the plot be facetted?}

\item{\code{fill}}{A vector of colors to use for filling posterior distributions.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalAnalyses-eval_models"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalAnalyses-eval_models}{}}}
\subsection{Method \code{eval_models()}}{
Compare multiple JAGS models using the WAIC and leave-one-out
cross-validation methods.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalAnalyses$eval_models(
  vars = "all",
  elev_span = 2500,
  excl_zone = 250,
  std_from = c("top", "bottom", "none")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{A character vector of variables to plot.}

\item{\code{elev_span}}{Length of elevation spans to use.}

\item{\code{excl_zone}}{Length of exclusion zones to use.}

\item{\code{std_from}}{Edge of elevational gradients to use. One of \code{top}, \code{bottom},
\code{none}. Use \code{none} for data that were not standardized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalAnalyses-get_statistical_details"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalAnalyses-get_statistical_details}{}}}
\subsection{Method \code{get_statistical_details()}}{
Compute Bayesian R-squared and the probability of posterior
fits to be lower than 0
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalAnalyses$get_statistical_details(
  vars = "all",
  elev_span = NULL,
  excl_zone = NULL,
  std_from = c("top", "bottom", "none")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{A character vector of variables to plot.}

\item{\code{elev_span}}{Length of elevation spans to use.}

\item{\code{excl_zone}}{Length of exclusion zones to use.}

\item{\code{std_from}}{Edge of elevational gradients to use. One of \code{top}, \code{bottom},
\code{none}. Use \code{none} for data that were not standardized.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GlobalAnalyses-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GlobalAnalyses-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GlobalAnalyses$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LocalAnalyses-new}{\code{LocalAnalyses$new()}}
\item \href{#method-LocalAnalyses-slope_histograms}{\code{LocalAnalyses$slope_histograms()}}
\item \href{#method-LocalAnalyses-influence_elev_span}{\code{LocalAnalyses$influence_elev_span()}}
\item \href{#method-LocalAnalyses-tbl_slope_summary}{\code{LocalAnalyses$tbl_slope_summary()}}
\item \href{#method-LocalAnalyses-clone}{\code{LocalAnalyses$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalAnalyses-new"></a>}}
\if{latex}{\out{\hypertarget{method-LocalAnalyses-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalAnalyses$new(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to a directory containing the outputs of the local-scale
analyses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalAnalyses-slope_histograms"></a>}}
\if{latex}{\out{\hypertarget{method-LocalAnalyses-slope_histograms}{}}}
\subsection{Method \code{slope_histograms()}}{
Plot histograms summarizing slope estimates in each location.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalAnalyses$slope_histograms(excl_zone = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{excl_zone}}{Size of exclusion zones to use. By default, make a
facetted plot using all exclusion zone sizes from the data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalAnalyses-influence_elev_span"></a>}}
\if{latex}{\out{\hypertarget{method-LocalAnalyses-influence_elev_span}{}}}
\subsection{Method \code{influence_elev_span()}}{
Plot the influence of elevation span on slope direction within
mountains.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalAnalyses$influence_elev_span(trs, excl_zone = 250)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trs}}{Dataset containing elevation spans.}

\item{\code{excl_zone}}{Size of the exclusion zone to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalAnalyses-tbl_slope_summary"></a>}}
\if{latex}{\out{\hypertarget{method-LocalAnalyses-tbl_slope_summary}{}}}
\subsection{Method \code{tbl_slope_summary()}}{
Summarize the proportion of positive and negative slopes based
on the level of uncertainty. Values are given in percentage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalAnalyses$tbl_slope_summary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LocalAnalyses-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LocalAnalyses-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocalAnalyses$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
