% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{make_regression_data}
\alias{make_regression_data}
\title{Prepare data for regression plots}
\usage{
make_regression_data(files, by_land_type = FALSE)
}
\arguments{
\item{files}{Files to get data from.}

\item{by_land_type}{Should estimates be calculated by land type?}
}
\description{
Structure data for \code{\link[=plot_regressions]{plot_regressions()}}.
}
\examples{
\dontrun{
get_files("path_to_dir", vars = c("dtr", "ts"), elevation_span = 2000) |>
  make_regression_data()
}
}
