% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{make_regression_data}
\alias{make_regression_data}
\title{Prepare data for regression plots}
\usage{
make_regression_data(files, type = c("draws", "land_types"))
}
\arguments{
\item{files}{Files to get data from.}

\item{type}{Type of data to make. One of \code{c(draws, land_types)}.}
}
\description{
Structure data for \code{\link[=regressions]{regressions()}}.
}
