% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressions.R
\name{plot_regressions}
\alias{plot_regressions}
\title{Plot regression fits}
\usage{
plot_regressions(
  data,
  facet_cols = "expl_var",
  facet_rows = "exclusion_zone",
  labels = NULL,
  ...
)
}
\arguments{
\item{data}{A list of data produced by \code{\link[=make_regression_data]{make_regression_data()}}.}

\item{facet_cols}{Variable defining faceting groups on the columns dimension.}

\item{facet_rows}{Variable defining faceting groups on the rows dimension.}

\item{labels}{A character vector used to label the x axis. If \code{labels} contains
several labels, each label must be named. Names must be the same as the
explanatory variables.}

\item{...}{Other arguments passed to methods. See details.}
}
\description{
Plot regression fits.
}
\details{
Other parameters for objects with S3 class \code{draws}:

\strong{n_draws} Number of draws to sample from the 95\% credible interval.

\strong{draws_prob} Probability mass to sample the draws from.

\strong{point_labels} Should point labels be shown?

\strong{seed} Seed value for draw sampling.
}
\examples{
\dontrun{
get_files("path_to_dir", vars = c("dtr", "ts"), elevation_span = 2000) |>
  make_regression_data() |>
  plot_regressions()
}
}
