% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis.R
\name{rs_reclass_dem}
\alias{rs_reclass_dem}
\title{Reclassify a DEM object}
\usage{
rs_reclass_dem(x, binwidth = 100, right = FALSE)
}
\arguments{
\item{x}{A raster object.}

\item{binwidth}{The width of the bins.}

\item{right}{Should bin intervals be closed on the right?
See \code{\link[terra:classify]{terra::classify()}} for details.}
}
\description{
Divide values of a DEM object into bins of a given width.
}
